Initial Design Complete
Modifications to the SparkFun design (most significant to least):
1) Rotate and relocate sensor as close as possible to edge of board.
Reason: I want to be able to put this sensor board into an enclosure
next to an LCD. The sensor needs to be as close to the LCD for both to
fit in the window.

2) Add location on back of board for large electrolytic capacitor.
Reason: placing the capacitor behind the sensor improves packaging
options for the board and sensor.

3) Added voltage regulator and level shifting components to make the
board 5V safe.
Reason: Easier to integrate with Arduinos and other 5V micros.
Note: In my understanding of the APDS datasheet the sensor can pull up
to 790mA. Because of this, a 1A regulator was selected, MCP1826T. Device
and footprint created in my Eagle library
(https://github.com/akreager/ATK_Engineering-Eagle_Library). I2C level
shifter design "borrowed" from SparkFun login level shifter board.

4) (Least important) Attempted to add provisions to be able to break
board into two pieces and reconnet with jumper wires.
Reason: Packaging.